Move world events to API, move some effects to API, make RandomEffect abstract, no implementations for it yet.
