+++ src/main/java/org/spout/vanilla/plugin/data/effect/type/MusicDiscEffect.java
@@ -31,7 +31,7 @@ import org.spout.api.geo.discrete.Point;
 
 import org.spout.vanilla.api.data.Music;
 
-import org.spout.vanilla.plugin.data.effect.GeneralEffect;
+import org.spout.vanilla.api.data.effect.GeneralEffect;
 
 public class MusicDiscEffect extends GeneralEffect {
 	private static final int MUSIC_RANGE = 48;
