+++ src/main/java/org/spout/vanilla/plugin/data/effect/type/NoteParticleEffect.java
@@ -32,7 +32,7 @@ import org.spout.api.geo.discrete.Point;
 
 import org.spout.vanilla.api.event.block.BlockActionEvent;
 
-import org.spout.vanilla.plugin.data.effect.GeneralEffect;
+import org.spout.vanilla.api.data.effect.GeneralEffect;
 import org.spout.vanilla.plugin.material.VanillaMaterials;
 
 public class NoteParticleEffect extends GeneralEffect {
